Fix a compilation warning in a WebKit2 GTK+ API test.
Reviewed by Gustavo Noronha Silva.

* UIProcess/API/gtk/tests/TestBackForwardList.cpp:
(testBackForwardListLimitAndCache): Use an int instead of size_t
because we want to print them as ints and do subtraction on them.

Canonical link: https://commits.webkit.org/94118@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@106145 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/API/gtk/tests/TestBackForwardList.cpp
@@ -27,7 +27,7 @@
 #include <webkit2/webkit2.h>
 
 // Back forward list limit is 100 by default.
-static const size_t kBackForwardListLimit = 100;
+static const int kBackForwardListLimit = 100;
 
 static WebKitTestServer* kServer;
 
@@ -240,7 +240,7 @@ static void testBackForwardListNavigation(BackForwardListTest* test, gconstpoint
 
 static void testBackForwardListLimitAndCache(BackForwardListTest* test, gconstpointer)
 {
-    for (size_t i = 0; i < kBackForwardListLimit; i++) {
+    for (int i = 0; i < kBackForwardListLimit; i++) {
         GOwnPtr<char> path(g_strdup_printf("/Page%d", i));
         test->m_changedFlags = BackForwardListTest::CurrentItem | BackForwardListTest::AddedItem;
         test->loadURI(kServer->getURIForPath(path.get()).data());
